fix: the Jetson example, and 84 clippy warnings in never-linted platform code
examples/gpu_control.rs called Simon::new() without `?`; it returns a Result.

The rest is `cargo clippy --fix` over the Linux platform code, which no lint had
ever seen: 102 warnings down to 18, all mechanical (unused imports and variables,
redundant closures, needless returns). CI runs clippy with -D warnings, so these
are build failures there, not advisories.

Windows, Linux, and macOS all compile; the eight test suites pass.

Eighteen warnings remain on Linux and are left for a separate pass: they include
dead-code findings (`body_has_array` never used, fields never read) that are
questions about whether the code should be deleted or wired up, not lint noise to
silence.

+++ src/agent/state.rs
@@ -211,7 +211,7 @@ impl SystemState {
                         .unwrap_or_else(|| "CPU".to_string()),
                     cores: num_cpus,
                     threads: num_cpus,
-                    utilization: utilization as f32,
+                    utilization: utilization,
                     temperature_c: None,
                     frequency_mhz: stats
                         .cores
@@ -220,7 +220,7 @@ impl SystemState {
                     per_core_usage: stats
                         .cores
                         .iter()
-                        .map(|c| (100.0 - c.idle.unwrap_or(100.0)) as f32)
+                        .map(|c| (100.0 - c.idle.unwrap_or(100.0)))
                         .collect(),
                 });
             }
